backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
pydicom>=2.4
numpy>=1.24
pillow>=10.0
h5py>=3.9
scikit-image>=0.22
openpyxl>=3.1

# TensorFlow (pick the right one for your platform)
tensorflow-macos>=2.13; platform_system == "Darwin" and platform_machine == "arm64"
tensorflow>=2.13; not (platform_system == "Darwin" and platform_machine == "arm64")


